data/data_collection/requirements.txt
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
distlib==0.3.9
filelock==3.12.2
google-cloud==0.34.0
idna==3.10
importlib-metadata==6.7.0
Pillow==9.5.0
pipenv==2023.10.3
platformdirs==4.0.0
requests==2.31.0
soupsieve==2.4.1
typing-extensions==4.7.1
urllib3==2.0.7
virtualenv==20.26.6
zipp==3.15.0
